Modify default options and pass them to modelStudio

This function returns default options for modelStudio. It is possible to modify values of this list and pass it to the options parameter in the main function. WARNING: Editing default options may cause unintended behavior.

modelStudioOptions(...)

Arguments

...
Options to change, option_name = value.

Value

list of options for modelStudio.

Main options:

scale_plot

TRUE Makes every plot the same height, ignores bar_width.

show_boxplot

TRUE Display boxplots in Feature Importance and Shapley Values plots.

show_subtitle

TRUE Should the subtitle be displayed?

subtitle

label parameter from explainer.

ms_title

Title of the dashboard.

margin_*

Plot margins. Change margin_left for longer/shorter axis labels.

w

420 in px. Inner plot width.

h

280 in px. Inner plot height.

bar_width

16 in px. Default width of bars for all plots, ignored when scale_plot = TRUE.

line_size

2 in px. Default width of lines for all plots.

point_size

3 in px. Default point radius for all plots.

[bar,line,point]_color

[#46bac2,#46bac2,#371ea3]

positive_color

#8bdcbe for Break Down and Shapley Values bars.

negative_color

#f05a71 for Break Down and Shapley Values bars.

default_color

#371ea3 for Break Down bar and highlighted line.

Plot specific options:

** is a two letter code unique to each plot, might be one of [bd,sv,cp,fi,pd,ad,fd,tv,at].

**_title

Plot specific title. Default varies.

**_subtitle

Plot specific subtitle. Default is subtitle.

**_bar_width

Plot specific width of bars. Default is bar_width, ignored when scale_plot = TRUE.

**_line_size

line_size Plot specific width of lines. Default is line_size.

**_point_size

Plot specific point radius. Default is point_size.

**_*_color

Plot specific [bar,line,point] color. Default is [bar,line,point]_color.

References

The input object is implemented in DALEX

Feature Importance, Ceteris Paribus, Partial Dependence and Accumulated Dependence plots are implemented in ingredients

Break Down and Shapley Values plots are implemented in iBreakDown

See also

Examples

library("DALEX")
library("modelStudio")

# fit a model
model_apartments <- glm(m2.price ~. , data = apartments)

# create an explainer for the model
explainer_apartments <- DALEX::explain(model_apartments,
                                       data = apartments,
                                       y = apartments$m2.price,
                                       label = "glm")

# pick observations
new_observation <- apartments[1:2,]
rownames(new_observation) <- c("ap1","ap2")

# modify default options
new_options <- modelStudioOptions(
  show_subtitle = TRUE,
  bd_subtitle = "Hello World",
  line_size = 5,
  point_size = 9,
  line_color = "pink",
  point_color = "purple",
  bd_positive_color = "yellow",
  bd_negative_color = "orange"
)

# make a studio for the model
modelStudio(explainer_apartments,
            new_observation,
            options = new_options,
            N = 200,  B = 5) # faster example