Get modified options for modelStudio

This function returns default options for modelStudio. It is possible to modify values of this list and pass it to options parameter in the main function. WARNING: Editing default options may cause unintended behavior.

modelStudioOptions(...)

Arguments

...
options to change, option_name = value.

Value

list of options for modelStudio

Main options:

scale_plot

TRUE Makes every plot the same height, ignores bar_width.

show_subtitle

TRUE Should subtitle be displayed?

subtitle

label parameter from explainer.

margin_*

plot margins. Change margin_left for longer/shorter axis labels.

w

420 in px. Inner plot width.

h

280 in px. Inner plot height.

bar_width

16 in px. Default width of bars for all plots, ignored when scale_plot = TRUE.

line_size

2 in px. Default width of lines for all plots.

point_size

3 in px. Default point radius for all plots.

[bar,line,point]_color

[#46bac2,#46bac2,#371ea3]

positive_color

#8bdcbe for Break Down and SHAP Values bars.

negative_color

#f05a71 for Break Down and SHAP Values bars.

default_color

#371ea3 for Break Down bar and highlighted line.

Plot specific options:

** is a two letter code unique to each plot, might be one of [bd,sv,cp,fi,pd,ad,fd].

**_title

plot specific title. Default varies.

**_subtitle

plot specific subtitle. Default is subtitle.

**_bar_width

plot specific width of bars. Default is bar_width, ignored when scale_plot = TRUE.

**_line_size

line_size plot specific width of lines. Default is line_size.

**_point_size

plot specific point radius. Default is point_size.

**_*_color

plot specific [bar,line,point] color. Default is [bar,line,point]_color.

Examples

library("modelStudio")

apartments <- DALEX::apartments

model_apartments <- glm(m2.price ~. ,
                        data = apartments)

explain_apartments <- DALEX::explain(model_apartments,
                                     data = apartments[,-1],
                                     y = apartments[,1],
                                     verbose = FALSE)

new_apartments <- apartments[1:2,]
rownames(new_apartments) <- c("ap1","ap2")

op <- modelStudioOptions(
  show_subtitle = TRUE,
  bd_subtitle = "Hello World",
  line_size = 5,
  point_size = 9,
  line_color = "pink",
  point_color = "purple",
  bd_positive_color = "yellow",
  bd_negative_color = "orange"
)

modelStudio(explain_apartments, new_apartments,
            facet_dim = c(1,2), N = 100, B = 10, show_info = FALSE,
            options = op)