Update the options of a modelStudio object

This function updates the options of a modelStudio object. WARNING: Editing default options may cause unintended behavior.

ms_update_options(object, ...)

Arguments

object

A modelStudio created with modelStudio().

...

Options to change in the form option_name = value, e.g. time = 0, facet_dim = c(1,2).

Value

An object of the r2d3, htmlwidget, modelStudio class.

Options

Main options:

scale_plot

TRUE Makes every plot the same height, ignores bar_width.

show_boxplot

TRUE Display boxplots in Feature Importance and Shapley Values plots.

show_subtitle

TRUE Should the subtitle be displayed?

subtitle

label parameter from explainer.

ms_title

Title of the dashboard.

ms_subtitle

Subtitle of the dashboard (makes space between the title and line).

ms_margin_*

Dashboard margins. Change margin_top for more ms_subtitle space.

margin_*

Plot margins. Change margin_left for longer/shorter axis labels.

w

420 in px. Inner plot width.

h

280 in px. Inner plot height.

bar_width

16 in px. Default width of bars for all plots, ignored when scale_plot = TRUE.

line_size

2 in px. Default width of lines for all plots.

point_size

3 in px. Default point radius for all plots.

[bar,line,point]_color

[#46bac2,#46bac2,#371ea3]

positive_color

#8bdcbe for Break Down and Shapley Values bars.

negative_color

#f05a71 for Break Down and Shapley Values bars.

default_color

#371ea3 for Break Down bar and highlighted line.

Plot-specific options:

** is a two letter code unique to each plot, might be one of [bd,sv,cp,fi,pd,ad,rv,fd,tv,at].

**_title

Plot-specific title. Default varies.

**_subtitle

Plot-specific subtitle. Default is subtitle.

**_axis_title

Plot-specific axis title. Default varies.

**_bar_width

Plot-specific width of bars. Default is bar_width, ignored when scale_plot = TRUE.

**_line_size

Plot-specific width of lines. Default is line_size.

**_point_size

Plot-specific point radius. Default is point_size.

**_*_color

Plot-specific [bar,line,point] color. Default is [bar,line,point]_color.

References

The input object is implemented in DALEX

Feature Importance, Ceteris Paribus, Partial Dependence and Accumulated Dependence explanations are implemented in ingredients

Break Down and Shapley Values explanations are implemented in iBreakDown

See also

Examples

library("DALEX")
library("modelStudio")

# fit a model
model_titanic <- glm(survived ~., data = titanic_imputed, family = "binomial")

# create an explainer for the model
explainer_titanic <- explain(model_titanic,
                             data = titanic_imputed,
                             y = titanic_imputed$survived)
#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  2207  rows  8  cols 
#>   -> target variable   :  2207  values 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  No value for predict function target column. (  default  )
#>   -> model_info        :  package stats , ver. 4.1.3 , task classification (  default  ) 
#>   -> predicted values  :  numerical, min =  0.008128381 , mean =  0.3221568 , max =  0.9731431  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9628583 , mean =  -2.569729e-10 , max =  0.9663346  
#>   A new explainer has been created!  

# make a studio for the model
ms <- modelStudio(explainer_titanic,
                  N = 200,  B = 5) # faster example
#> `new_observation` argument is NULL. `new_observation_n` observations needed to calculate local explanations are taken from the data.

# update the options
new_ms <- ms_update_options(ms,
                            time = 0,
                            facet_dim = c(1,2),
                            margin_left = 150)
new_ms